+++ packages/amgi-aiobotocore/tests_amgi_aiobotocore/test_sqs_message_integration.py
@@ -17,7 +17,7 @@ class _StrMatcher:
         return isinstance(other, str)
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 async def localstack_container() -> AsyncGenerator[LocalStackContainer, None]:
     with LocalStackContainer(
         image="localstack/localstack:4.9.2"
@@ -67,20 +67,9 @@ async def app(
     )
     loop = asyncio.get_running_loop()
     serve_task = loop.create_task(server.serve())
-    async with app.call() as (scope, receive, send):
-        assert scope == {
-            "amgi": {"spec_version": "1.0", "version": "1.0"},
-            "type": "lifespan",
-            "state": {},
-        }
-        lifespan_startup = await receive()
-        assert lifespan_startup == {"type": "lifespan.startup"}
-        await send({"type": "lifespan.startup.complete"})
+    async with app.lifespan():
         yield app
         server.stop()
-        lifespan_shutdown = await receive()
-        assert lifespan_shutdown == {"type": "lifespan.shutdown"}
-        await send({"type": "lifespan.shutdown.complete"})
 
     await serve_task